Command and Message Managers
This commit created the 'CommandManager' and 'MessageManager' classes in
the main package and created methods for using these classes. I also
made some modifications to the main class to coincide with this and in a
new class called 'commands' added our first command (the /guild create
command) and made the 'SubCommand' structure class.